cleaned up form by refactoring and outsourcing components, added additional validation functions and added a success page  when submitting the form
